make events query work
