+++ posthog/api/query.py
@@ -11,6 +11,7 @@ from rest_framework.request import Request
 from posthog.api.documentation import extend_schema
 from posthog.api.routing import StructuredViewSetMixin
 from posthog.exceptions import RequestParsingError
+from posthog.models import Team
 from posthog.models.event.query_event_list import query_events_list_v2
 from posthog.permissions import ProjectMembershipNecessaryPermissions, TeamMemberAccessPermission
 from posthog.rate_limit import PassThroughClickHouseBurstRateThrottle, PassThroughClickHouseSustainedRateThrottle
@@ -32,31 +33,13 @@ class QueryViewSet(StructuredViewSetMixin, viewsets.ViewSet):
     )
     def list(self, request: Request, **kw) -> HttpResponse:
         query_json = self._query_json_from_request(request)
-        query_result = self.process_query(query_json)
+        query_result = process_query(self.team, query_json)
         return JsonResponse(query_result)
 
     def post(self, request, *args, **kwargs):
         query_json = request.data
-        query_result = self.process_query(query_json)
+        query_result = process_query(self.team, query_json)
         return JsonResponse(query_result)
-
-    def process_query(self, query_json: Dict) -> Dict:
-        query_kind = query_json.get("kind")
-        if query_kind == "EventsQuery":
-            query = EventsQuery.parse_obj(query_json)
-            query_result = query_events_list_v2(
-                team=self.team,
-                query=query,
-            )
-            # :KLUDGE: Calling `query_result.dict()` without the following deconstruction fails with a cryptic error
-            return {
-                "columns": query_result.columns,
-                "types": query_result.types,
-                "results": query_result.results,
-                "hasMore": query_result.hasMore,
-            }
-        else:
-            raise RequestParsingError("Unsupported query kind: %s" % query_kind)
 
     def _query_json_from_request(self, request):
         if request.method == "POST":
@@ -78,3 +61,22 @@ class QueryViewSet(StructuredViewSetMixin, viewsets.ViewSet):
         except (json.JSONDecodeError, UnicodeDecodeError) as error_main:
             raise RequestParsingError("Invalid JSON: %s" % (str(error_main)))
         return query
+
+
+def process_query(team: Team, query_json: Dict) -> Dict:
+    query_kind = query_json.get("kind")
+    if query_kind == "EventsQuery":
+        query = EventsQuery.parse_obj(query_json)
+        query_result = query_events_list_v2(
+            team=team,
+            query=query,
+        )
+        # :KLUDGE: Calling `query_result.dict()` without the following deconstruction fails with a cryptic error
+        return {
+            "columns": query_result.columns,
+            "types": query_result.types,
+            "results": query_result.results,
+            "hasMore": query_result.hasMore,
+        }
+    else:
+        raise RequestParsingError("Unsupported query kind: %s" % query_kind)
