+++ posthog/api/test/test_query.py
@@ -1,0 +1,126 @@
+from freezegun import freeze_time
+
+from posthog.schema import AnyPropertyFilter, EventsQuery
+from posthog.test.base import (
+    APIBaseTest,
+    ClickhouseTestMixin,
+    _create_event,
+    _create_person,
+    also_test_with_materialized_columns,
+    flush_persons_and_events,
+    snapshot_clickhouse_queries,
+)
+
+
+class TestQuery(ClickhouseTestMixin, APIBaseTest):
+    ENDPOINT = "query"
+
+    @snapshot_clickhouse_queries
+    def test_select_hogql_expressions(self):
+        with freeze_time("2020-01-10 12:00:00"):
+            _create_person(
+                properties={"email": "tom@posthog.com"},
+                distinct_ids=["2", "some-random-uid"],
+                team=self.team,
+                immediate=True,
+            )
+            _create_event(team=self.team, event="sign up", distinct_id="2", properties={"key": "test_val1"})
+        with freeze_time("2020-01-10 12:11:00"):
+            _create_event(team=self.team, event="sign out", distinct_id="2", properties={"key": "test_val2"})
+        with freeze_time("2020-01-10 12:12:00"):
+            _create_event(team=self.team, event="sign out", distinct_id="3", properties={"key": "test_val2"})
+        with freeze_time("2020-01-10 12:13:00"):
+            _create_event(team=self.team, event="sign out", distinct_id="4", properties={"key": "test_val3"})
+        flush_persons_and_events()
+
+        with freeze_time("2020-01-10 12:14:00"):
+            query = EventsQuery(select=["event", "distinct_id", "properties.key", "concat(event, ' ', properties.key)"])
+            response = self.client.post(f"/api/projects/{self.team.id}/query/", query.dict()).json()
+            self.assertEqual(
+                response,
+                {
+                    "columns": ["event", "distinct_id", "properties.key", "concat(event, ' ', properties.key)"],
+                    "hasMore": False,
+                    "types": ["String", "String", "String", "String"],
+                    "results": [
+                        ["sign out", "4", "test_val3", "sign out test_val3"],
+                        ["sign out", "3", "test_val2", "sign out test_val2"],
+                        ["sign out", "2", "test_val2", "sign out test_val2"],
+                        ["sign up", "2", "test_val1", "sign up test_val1"],
+                    ],
+                },
+            )
+
+            query.select = ["*", "event"]
+            response = self.client.post(f"/api/projects/{self.team.id}/query/", query.dict()).json()
+            self.assertEqual(response["columns"], ["*", "event"])
+            self.assertIn("Tuple(", response["types"][0])
+            self.assertEqual(response["types"][1], "String")
+            self.assertEqual(len(response["results"]), 4)
+            self.assertIsInstance(response["results"][0][0], dict)
+            self.assertIsInstance(response["results"][0][1], str)
+
+            query.select = ["count()", "event"]
+            response = self.client.post(f"/api/projects/{self.team.id}/query/", query.dict()).json()
+            self.assertEqual(
+                response,
+                {
+                    "columns": ["count()", "event"],
+                    "hasMore": False,
+                    "types": ["UInt64", "String"],
+                    "results": [[3, "sign out"], [1, "sign up"]],
+                },
+            )
+
+            query.select = ["count()", "event"]
+            query.where = ['event == "sign up" or like(properties.key, "%val2")']
+            query.orderBy = ["-count()", "event"]
+            response = self.client.post(f"/api/projects/{self.team.id}/query/", query.dict()).json()
+            self.assertEqual(
+                response,
+                {
+                    "columns": ["count()", "event"],
+                    "hasMore": False,
+                    "types": ["UInt64", "String"],
+                    "results": [[2, "sign out"], [1, "sign up"]],
+                },
+            )
+
+    @also_test_with_materialized_columns(["key"])
+    @snapshot_clickhouse_queries
+    def test_hogql_property_filter(self):
+        with freeze_time("2020-01-10 12:00:00"):
+            _create_person(
+                properties={"email": "tom@posthog.com"},
+                distinct_ids=["2", "some-random-uid"],
+                team=self.team,
+                immediate=True,
+            )
+            _create_event(team=self.team, event="sign up", distinct_id="2", properties={"key": "test_val1"})
+        with freeze_time("2020-01-10 12:11:00"):
+            _create_event(team=self.team, event="sign out", distinct_id="2", properties={"key": "test_val2"})
+        with freeze_time("2020-01-10 12:12:00"):
+            _create_event(team=self.team, event="sign out", distinct_id="3", properties={"key": "test_val2"})
+        with freeze_time("2020-01-10 12:13:00"):
+            _create_event(team=self.team, event="sign out", distinct_id="4", properties={"key": "test_val3"})
+        flush_persons_and_events()
+
+        with freeze_time("2020-01-10 12:14:00"):
+            query = EventsQuery(
+                select=["event", "distinct_id", "properties.key", "'a%sd'", "concat(event, ' ', properties.key)"]
+            )
+
+            response = self.client.post(f"/api/projects/{self.team.id}/query/", query.dict()).json()
+            self.assertEqual(len(response["results"]), 4)
+
+            query.properties = [AnyPropertyFilter.parse_obj({"type": "hogql", "key": "'a%sd' == 'foo'"})]
+            response = self.client.post(f"/api/projects/{self.team.id}/query/", query.dict()).json()
+            self.assertEqual(len(response["results"]), 0)
+
+            query.properties = [AnyPropertyFilter.parse_obj({"type": "hogql", "key": "'a%sd' == 'a%sd'"})]
+            response = self.client.post(f"/api/projects/{self.team.id}/query/", query.dict()).json()
+            self.assertEqual(len(response["results"]), 4)
+
+            query.properties = [AnyPropertyFilter.parse_obj({"type": "hogql", "key": "properties.key == 'test_val2'"})]
+            response = self.client.post(f"/api/projects/{self.team.id}/query/", query.dict()).json()
+            self.assertEqual(len(response["results"]), 2)
