+++ posthog/schema.py
@@ -1,6 +1,6 @@
 # generated by datamodel-codegen:
 #   filename:  schema.json
-#   timestamp: 2023-01-26T12:59:55+00:00
+#   timestamp: 2023-01-26T14:08:14+00:00
 
 from __future__ import annotations
 
@@ -553,7 +553,7 @@ class EventsQuery(BaseModel):
     class Config:
         extra = Extra.forbid
 
-    actionId: Optional[float] = Field(None, description="Show events matching a given action")
+    actionId: Optional[int] = Field(None, description="Show events matching a given action")
     after: Optional[str] = Field(None, description="Only fetch events that happened after this timestamp")
     before: Optional[str] = Field(None, description="Only fetch events that happened before this timestamp")
     event: Optional[str] = Field(None, description="Limit to events matching this string")
@@ -562,8 +562,8 @@ class EventsQuery(BaseModel):
         description="Fixed properties in the query, can't be edited in the interface (e.g. scoping down by person)",
     )
     kind: str = Field("EventsQuery", const=True)
-    limit: Optional[float] = Field(None, description="Number of rows to return")
-    offset: Optional[float] = Field(None, description="Number of rows to skip before returning rows")
+    limit: Optional[int] = Field(None, description="Number of rows to return")
+    offset: Optional[int] = Field(None, description="Number of rows to skip before returning rows")
     orderBy: Optional[List[str]] = Field(None, description="Columns to order by")
     personId: Optional[str] = Field(None, description="Show events for a given person")
     properties: Optional[List[AnyPropertyFilter]] = Field(None, description="Properties configurable in the interface")
